+++ magentoerpconnect/CHANGES.rst
@@ -16,6 +16,7 @@ Changelog
 * Fix: calculate correctly the discount ratio on sales order lines (lp:1201590)
 * Possibility to exclude products from stock synchronizations
 * Products disabled on Magento are imported disabled on OpenERP. An additional module allows more options.
+* Possibility to disable import of sales orders per storeview
 
 
 2.3.1 (2014-01-23)